Sirket_Tekillestirme/company-resolver/resolver/requirements.txt
google-generativeai==0.8.4
numpy<2
torch
sentence_transformers
